requirements.txt
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-Session==0.8.0

python-dotenv==1.0.1
requests==2.32.3
scikit-learn==1.3.2
transformers==4.40.1
safetensors==0.4.3
torch==2.1.2
torchvision==0.16.2